Add player-wall collision

+++ Pacman_SDL/Vector2.h
@@ -1,11 +1,12 @@
 #pragma once
+
 class Vector2
 {
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 
 	Vector2();
-	Vector2(int x, int y);
+	Vector2(float x, float y);
 };
